+++ lib/shrine.rb
@@ -322,7 +322,8 @@ class Shrine
 
         # Does the actual uploading, calling `#upload` on the storage.
         def copy(io, context)
-          storage.upload(io, context[:location], context[:metadata])
+          context[:upload_options] = (context[:upload_options] || {}).merge(shrine_metadata: context[:metadata])
+          storage.upload(io, context[:location], context[:upload_options])
         ensure
           io.close rescue nil
         end