+++ test/s3_test.rb
@@ -61,7 +61,7 @@ describe Shrine::Storage::S3 do
     end
 
     it "accepts additional upload options via metadata" do
-      @s3.upload(fakeio, "foo", {"s3" => {content_type: "foo/bar"}})
+      @s3.upload(fakeio, "foo", content_type: "foo/bar")
       tempfile = @s3.download("foo")
       assert_equal "foo/bar", tempfile.content_type
     end